==> notebooks/data retrieval/red links sample unique/02. Manage redirections.ipynb <==
import pandas as pd

PATH_WIKI = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'
data_time = '20180920'

df_articles = pd.read_csv(PATH_WIKI+"ukwiki-" + data_time + "-pages-meta-current.xml_red.csv",
                          encoding='UTF-8', quotechar="\"")
df_articles.head()

df_redirect1 = df_articles[['title']]
df_redirect1.columns = ['redirect']

df_redirect1.head()

df_redirect2 = df_articles[['redirect']]

df_redirect2.head()

frames = [df_redirect1, df_redirect2]
df_all_redirections = pd.concat(frames)

df_all_redirections.head()

df_all_redirections = df_all_redirections.drop_duplicates()

df_all_redirections.shape

df_all_redirections.to_csv(PATH_WIKI+"ukwiki-" + data_time + "-redirections.csv", 
                     index = False, encoding='UTF-8', quotechar="\"", sep ="^")

# results in a file "ukwiki-" + data_time + "-redirections.csv"
