==> notebooks/data retrieval/red links sample unique/03. Get all article names.ipynb <==
import os
import pandas as pd
import re

PATH_WIKI = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'
data_time = '20180920'

## Get all article names

UKWIKI_ART_FNMS = []
for file in os.listdir(PATH_WIKI):
    if re.match(r"ukwiki-" + data_time + "-pages-meta-current.xml_art.csv", file):
        UKWIKI_ART_FNMS.append(file)
print('UKWIKI_ART_FNMS:', UKWIKI_ART_FNMS)

df_uk_id_name = None
for fn in UKWIKI_ART_FNMS:
    fn = PATH_WIKI+fn
    print(fn)
    df_articles = pd.read_csv(fn, encoding='UTF-8', quotechar="\"")
    print(df_articles.head(100))
    df_id_name_tmp = df_articles[['id', 'title', 'text_len']].drop_duplicates()
    df_id_name_tmp.columns = ['id', 'title', 'length']
    #print("df_id_name_tmp size: {}".format(df_id_name_tmp.shape))
    if df_uk_id_name is not None:
        df_uk_id_name = df_uk_id_name.append(df_id_name_tmp)        
        #print("append")
    else:
        df_uk_id_name = df_id_name_tmp
        #print("assign")
    print("df_uk_id_name size: {}".format(df_uk_id_name.shape))
df_uk_id_name = df_uk_id_name.drop_duplicates()

df_uk_id_name.to_csv(PATH_WIKI+"ukwiki-" + data_time + "-id_name.csv", 
                     index = False, encoding='UTF-8', quotechar="\"", sep ="^")

# results in a file "ukwiki-" + data_time + "-id_name.csv"
