==> notebooks/data retrieval/red links sample unique/04. Get links between all pages.ipynb <==
import pandas as pd

PATH_WIKI = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'
data_time = '20180920'

df_uk_redirections = pd.read_csv(PATH_WIKI+"ukwiki-" + data_time + "-redirections.csv", 
                            encoding='UTF-8', quotechar="\"", sep ="^")
df_uk_redirections.columns = ['title']
df_uk_redirections.head()

print(df_uk_redirections.shape)

df_uk_id_name = pd.read_csv(PATH_WIKI+"ukwiki-" + data_time + "-id_name.csv", 
                         encoding='UTF-8', quotechar="\"", sep ="^")

df_uk_name = df_uk_id_name[['title']]
df_uk_name.head()

frames = [df_uk_redirections, df_uk_name]
df_uk_name_all = pd.concat(frames)

df_uk_name_all['title'] = df_uk_name_all['title'].str.strip()

df_uk_name_all.head()

df_uk_name_all_strip = df_uk_name_all.iloc[:,0]

df_uk_name_all_strip = df_uk_name_all_strip.str.replace('  ', ' ')

df_uk_name_all_stripped = df_uk_name_all_strip.to_frame()

df_uk_name_all_stripped.head()

df_uk_name_all_stripped = df_uk_name_all_stripped.drop_duplicates()

df_uk_name_all.shape

df_uk_name_all['title_lc'] = df_uk_name_all['title'].str.lower()

df_uk_name_all.head()

fn = PATH_WIKI+"ukwiki-" + data_time + "-pages-meta-current.xml_art.csv"
print(fn)
df_articles = pd.read_csv(fn, encoding='UTF-8', quotechar="\"")
df_articles['link_val'] = df_articles['link_val'].str.split("#").str[0] # ???
df_articles = df_articles[df_articles['link_val'].str.strip() != ''] 

df_articles.head()

df_articles['title'] = df_articles['title'].str.strip()
df_articles['link_val'] = df_articles['link_val'].str.strip()

df_uk_titles_series = df_articles.iloc[:,4]

df_uk_titles_series_stripped = df_uk_titles_series.str.replace('  ', ' ')

df_titles_stripped = df_uk_titles_series_stripped.to_frame()

df_titles_stripped.head()

df_titles_stripped['link_val'] = df_titles_stripped['link_val'].str.lower()

print(len(df_titles_stripped))
print(len(df_articles))

# Create files with page links

df_articles_with_lc = pd.concat([df_articles, df_titles_stripped], axis=1)

df_articles_with_lc.head()

df_uk_id_name = pd.read_csv(PATH_WIKI+"ukwiki-" + data_time + "-id_name.csv", 
                         encoding='UTF-8', quotechar="\"", sep ="^")

df_uk_id_name = df_uk_id_name[['id', 'title']]

df_uk_id_name.head()

redirections = PATH_WIKI+'ukwiki-' + data_time + '-pages-meta-current.xml_red.csv'

df_uk_redirections = pd.read_csv(redirections, encoding='UTF-8', quotechar="\"")

df_uk_redirections = df_uk_redirections1[['id', 'title']]

df_uk_redirections.head()

df_uk_id_name.head()

df_uk_redirections.head()

df_all_ids_uk = pd.concat([df_uk_id_name, df_uk_redirections], axis=0)

df_all_ids_uk.head()

df_all_ids_uk.shape

df_all_ids_uk['title'] = df_all_ids_uk['title'].str.strip()

df_all_ids_uk_series = df_all_ids_uk.iloc[:,1]

df_all_ids_uk_series_stripped = df_all_ids_uk_series.str.replace('  ', ' ')
df_all_ids_uk_stripped = df_all_ids_uk_series_stripped.to_frame()

df_all_ids_uk_stripped['title'] = df_all_ids_uk_stripped['title'].str.lower()

df_all_ids_uk_stripped.shape

df_all_ids_uk_cleaned = pd.concat([df_all_ids_uk, df_all_ids_uk_stripped], axis=1)

df_all_ids_uk_cleaned.columns = ['id', 'title', 'title_lc']

df_all_ids_uk_cleaned.head()

df_articles_with_lc.columns = ['id', 'title', 'text_len', 'link_pos', 'link_val', 'link_txt', 'link_val_lc']

df_articles_with_lc.head()

page_links_dirty = pd.merge(df_articles_with_lc, df_all_ids_uk_cleaned, how='left', left_on=['link_val_lc'], right_on=['title_lc'])

page_links_dirty = page_links_dirty[pd.notnull(page_links_dirty['link_val'])]

page_links_dirty.head()

page_links_with_titles = page_links_dirty[['id_x', 'title_x', 'id_y', 'link_val', 'title_lc']]

page_links_with_titles.head()

page_links_with_titles.to_csv(PATH_WIKI+'page_links_with_titles.csv', index = False, encoding='UTF-8', sep='^')

page_links_ids = page_links_with_titles[['id_x', 'id_y']]
page_links_ids.columns = ['id', 'link_id']

page_links_ids.to_csv(PATH_WIKI+'uk_page_links_ids.csv', index = False, encoding='UTF-8', sep='^')

# find red links among all links
links_which_exist = pd.merge(df_uk_name_all, df_titles_stripped,
                             how='right', left_on=['title_lc'], right_on=['link_val'])

links_which_exist = links_which_exist.drop_duplicates()

links_which_exist = links_which_exist.dropna(how='all')

links_which_exist

candidates_for_redlinks = links_which_exist[links_which_exist['title'].isnull()]

candidates_for_redlinks = candidates_for_redlinks[['link_val']]

candidates_for_redlinks.shape

candidates_for_redlinks.head()

df_articles.head()

df_uk_titles_series = df_articles.iloc[:,4]

df_uk_titles_series_stripped = df_uk_titles_series.str.replace('  ', ' ')

df_titles_stripped = df_uk_titles_series_stripped.to_frame()

df_titles_stripped['link_val_lc'] = df_titles_stripped['link_val'].str.lower()

df_pagelinks_with_lowercase = pd.concat([df_articles, df_titles_stripped], axis=1)

df_pagelinks_with_lowercase_redlinks = pd.merge(df_pagelinks_with_lowercase, candidates_for_redlinks,
         how='left', left_on=['link_val_lc'], right_on=['link_val'])

df_pagelinks_with_lowercase_redlinks

redlinks_with_all = df_pagelinks_with_lowercase_redlinks[~(df_pagelinks_with_lowercase_redlinks['link_val_y'].isnull())]

redlinks = redlinks_with_all[['id', 'link_val_x']]
redlinks.columns = ['id', 'link_val', 'dd']

redlinks = redlinks[['id', 'link_val']]

redlinks.to_csv(PATH_WIKI+'ukwiki-20180920-all_reslinks.csv', index = False, encoding='UTF-8', quotechar="\"")

df_articles.to_csv(PATH_WIKI+'ukwiki-20180920-pages-links.csv', index = False, encoding='UTF-8', quotechar="\"")

# results in files:
# 1. 'page_links_with_titles.csv',
# 2. 'uk_page_links_ids.csv',
# 3. 'ukwiki-20180920-all_reslinks.csv',
# 4. 'ukwiki-20180920-pages-links.csv.gz'
