==> notebooks/data retrieval/red links sample unique/06. Get red links sample.ipynb <==
import pandas as pd
import numpy as np

PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'
data_time = '20180920'

## Get red links sample

### Files needed for final processing:
### 1. 'ukwiki-20180920-pages-links.csv'
### 2. "ukwiki-" + data_time + "-id_name.csv"
### 3. '20180920-langlinks_uk_en.csv'
### 4. 'ukwiki-20180920-all_reslinks.csv'

df_uk_full = pd.read_csv(PATH_TO_DATA+'ukwiki-20180920-pages-links.csv')
print(df_uk_full.shape)
df_uk_full.head()

df_uk_red = pd.read_csv(PATH_TO_DATA+'ukwiki-20180920-all_reslinks.csv', encoding='UTF-8', quotechar="\"")

df_uk_red.head()

# article names
df_uk_name = pd.read_csv(PATH_TO_DATA+"ukwiki-" + data_time + "-id_name.csv", encoding='utf-8', sep="^")

print(df_uk_name.shape)
print(df_uk_name.columns)
df_uk_name.head()

# find maximum id among blue links
max_id = np.max(df_uk_name['id'])

# find unique titles among red links, create df with

red_links = df_uk_red['link_val'].unique()

red_links_ids = pd.DataFrame({'title': red_links, 
                               'id': np.arange(max_id+1, len(red_links)+max_id+1)})

red_links_ids.head()

df_uk_red_ext = df_uk_red.merge(right = red_links_ids, left_on = 'link_val', right_on = 'title', how = 'right')
df_uk_red_ext = df_uk_red_ext[['id_x', 'link_val', 'id_y']]
df_uk_red_ext.columns = ['id_parent', 'red_link_name', 'red_link_id']
print(df_uk_red_ext.shape)
df_uk_red_ext.head()

df_uk_red_ext.to_csv(PATH_TO_DATA+'all_uk_redlinks_with_parent_ids.csv', header=True, index=False)

# Load uk-en correpondences
fn = PATH_TO_DATA+'20180920-langlinks_uk_en.csv'
df_link = pd.read_csv(fn,  encoding='UTF-8', quotechar='\'')
print(df_link.shape)
print(df_link.columns)
df_link.head()

df_uk_red_ext = pd.read_csv(PATH_TO_DATA+'all_uk_redlinks_with_parent_ids.csv')
print(df_uk_red_ext.shape)
df_uk_red_ext.head()

# add en correspondences to uk parent pages for red links
df_uk_red_corresp = df_uk_red_ext.merge(right = df_link, left_on = 'id_parent', right_on = 'll_from', how = 'left')
df_uk_red_corresp.head()

df_uk_red_corresp = df_uk_red_corresp[pd.notnull(df_uk_red_corresp['ll_from'])]
df_uk_red_corresp.head()

df_uk_red_corresp = df_uk_red_corresp[['id_parent','red_link_name', 'red_link_id', 'll_title']]

df_uk_red_corresp.to_csv(PATH_TO_DATA+'df_allukredlinks_withentitle.csv', header=True, index=False)

df_uk_red_corresp = df_uk_red_corresp[['id_parent','red_link_name', 'red_link_id']]
df_uk_red_corresp = df_uk_red_corresp.sort_values(by = 'red_link_id')
df_uk_red_corresp = df_uk_red_corresp.drop_duplicates(keep = 'first')
df_uk_red_corresp = df_uk_red_corresp.dropna()
df_uk_red_corresp

# for every red link calculate number of incoming links
df_red_links_most_common = df_uk_red_corresp['red_link_id'].value_counts().reset_index()
df_red_links_most_common.head()

df_red_links_most_common.columns = ['red_link_id_', 'count']
df_red_links_most_common.head()

df_red_links_most_common = df_red_links_most_common[(df_red_links_most_common['count'] >= 25)]
df_red_links_most_common

result = df_uk_red_corresp.merge(right = df_red_links_most_common,
                                 left_on = 'red_link_id', right_on = 'red_link_id_', how = 'inner')

result

result_red_links_uk = result['red_link_name'].unique()
df_result_red_links_uk = pd.DataFrame(result_red_links_uk, columns=['red_link_name'])

# save red links to file
df_result_red_links_uk.to_csv(PATH_TO_DATA+'df_sample_red_links_uk.csv', header=True, index=False)

# results in files:
# 1. 'all_uk_redlinks_with_parent_ids.csv',
# 2. 'df_allukredlinks_withentitle.csv',
# 3. 'df_sample_red_links_uk.csv'
